Add wasm target to korge gradle plugin
